refactor: share the cached session-scan machinery
Extract the byte-identical file loader, fold helpers, and cached file
scanner into scan::compact, driven by a CompactSink each feature
implements. usage and analysis stop carrying parallel copies.
